update for dynamic port
Sonic Pi 3.2dev uses a dynamically allocated port to listen to incoming OSC messages. The file is adapted to read the allocated port from the Sonic Pi server-putput.log and to use the value to initialise sonic-pi-cli. If there is no port detected in the log it reverts to using the previous static value 4557 used in previous versions of Sonic Pi