Dynamic Port Access Without Use of CLI
This change modifies the `SonicPi` class so that it will find the
dynamic port that Sonic Pi is running on. By doing this, you can
instantiate `SonicPi` in irb or other ruby code.

The existing change to find the dynamic port would execute when running
`sonic_pi` from the command line tool; this extends that availability to
be wherever the `SonicPi` class can be used.